Reinstate asn1crypto registration of AA types
Some extension registrations were inadvertently removed in the process
of upgrading asn1crypto and the removal of associated monkey patches
(see 87aaf7f). This issue went undetected in the tests because our test
suite imports pyhanko_certvalidator, which performs the same
registrations on import.

This commit restores the registration hook and makes sure it gets called
when the CLI is loaded.

Fixes #4.